REFACTOR: Moved project related functions to project directory
